# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Stayinn API",
    "version": "1.0.0",
    "description": "API Specification for Stayinn — Hotel Booking Marketplace"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local Development Server"
    },
    {
      "url": "https://stayinn.ng/api",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "tags": ["Auth"],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fullName": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "phone": { "type": "string", "example": "+2348036930772" },
                  "password": { "type": "string" },
                  "role": { "type": "string", "enum": ["BOOKER", "HOTEL_MANAGER"] }
                },
                "required": ["fullName", "email", "phone", "password", "role"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Account created successfully" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login user",
        "tags": ["Auth"],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "identifier": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["identifier", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful" }
        }
      }
    },
...
    "/bookings": {
      "post": {
        "summary": "Create a booking",
        "tags": ["Bookings"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "hotelId": { "type": "string" },
                  "roomTypeId": { "type": "string" },
                  "checkInDate": { "type": "string", "format": "date" },
                  "checkOutDate": { "type": "string", "format": "date" },
                  "numberOfGuests": { "type": "integer" }
                },
                "required": ["hotelId", "roomTypeId", "checkInDate", "checkOutDate", "numberOfGuests"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Booking created" }
        }
      }
    },
...
    "/bookings/hotel/{hotelId}": {
      "get": {
        "summary": "Get bookings for a specific hotel (MANAGER/ADMIN)",
        "tags": ["Bookings"],
        "parameters": [
          { "name": "hotelId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Hotel's bookings" }
        }
      }
    },
    "/bookings/{id}": {
      "get": {
        "summary": "Get booking details by ID",
        "tags": ["Bookings"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Booking details" }
        }
      }
    },
    "/bookings/{id}/confirm": {
      "post": {
        "summary": "Confirm a booking via QR scan",
        "tags": ["Bookings"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "qrData": { "type": "string" },
                  "qrSignature": { "type": "string" }
                },
                "required": ["qrData", "qrSignature"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Booking confirmed" }
        }
      }
    },
    "/bookings/{id}/cancel": {
      "post": {
        "summary": "Cancel a booking",
        "tags": ["Bookings"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Booking cancelled" }
        }
      }
    },
...
  }
}