# src/app/api/auth/login/route.ts
import { NextResponse } from 'next/server';
import { prisma } from '@/lib/prisma';
import { comparePassword, signAccessToken, signRefreshToken } from '@/lib/auth';
import { loginSchema } from '@/lib/validation/auth';
import { rateLimit } from '@/lib/rate-limit';

export async function POST(req: Request) {
  try {
    const ip = req.headers.get('x-forwarded-for') || '127.0.0.1';
    const limit = await rateLimit(`auth:${ip}`, 5, 60);
    
    if (!limit.success) {
      return NextResponse.json({ success: false, error: { code: 'RATE_LIMITED', message: 'Too many requests' } }, { status: 429 });
    }

    const body = await req.json();
    const result = loginSchema.safeParse(body);
    
    if (!result.success) {
      return NextResponse.json({ success: false, error: { code: 'VALIDATION_ERROR', message: 'Invalid credentials' } }, { status: 422 });
    }

    const { identifier, password } = result.data;

    const user = await prisma.user.findFirst({
      where: {
        OR: [
          { email: identifier },
          { phone: identifier }
        ]
      }
    });

    if (!user) {
      return NextResponse.json({ success: false, error: { code: 'INVALID_CREDENTIALS', message: 'Wrong identifier or password' } }, { status: 401 });
    }

    const isMatch = await comparePassword(password, user.passwordHash);
    if (!isMatch) {
      return NextResponse.json({ success: false, error: { code: 'INVALID_CREDENTIALS', message: 'Wrong identifier or password' } }, { status: 401 });
    }

    const accessToken = await signAccessToken({ userId: user.id, email: user.email, role: user.role });
    const refreshToken = await signRefreshToken({ userId: user.id });

    return NextResponse.json({
      success: true,
      data: {
        user: { id: user.id, fullName: user.fullName, email: user.email, phone: user.phone, role: user.role },
        accessToken,
        refreshToken
      }
    });

  } catch (error: unknown) {
    console.error('Login Error:', error);
    return NextResponse.json({ success: false, error: { code: 'INTERNAL_ERROR', message: 'Something went wrong' } }, { status: 500 });
  }
}