# src/app/api/auth/register/route.ts
import { NextResponse } from 'next/server';
import { prisma } from '@/lib/prisma';
import { hashPassword, signAccessToken, signRefreshToken } from '@/lib/auth';
import { registerSchema } from '@/lib/validation/auth';
import { rateLimit } from '@/lib/rate-limit';
import { Role } from '@prisma/client';

export async function POST(req: Request) {
  try {
    const ip = req.headers.get('x-forwarded-for') || '127.0.0.1';
    const limit = await rateLimit(`auth:${ip}`, 5, 60);

    if (!limit.success) {
      return NextResponse.json({ success: false, error: { code: 'RATE_LIMITED', message: 'Too many requests' } }, { status: 429, headers: { 'X-RateLimit-Remaining': '0' } });
    }

    const body = await req.json();
    const result = registerSchema.safeParse(body);

    if (!result.success) {
      return NextResponse.json({ success: false, error: { code: 'VALIDATION_ERROR', message: result.error.issues[0].message } }, { status: 422 });
    }

    const { fullName, email, phone, password, role } = result.data;

    const existingEmail = await prisma.user.findUnique({ where: { email } });
    if (existingEmail) {
      return NextResponse.json({ success: false, error: { code: 'EMAIL_EXISTS', message: 'Email already registered' } }, { status: 409 });
    }

    const existingPhone = await prisma.user.findUnique({ where: { phone } });
    if (existingPhone) {
      return NextResponse.json({ success: false, error: { code: 'PHONE_EXISTS', message: 'Phone number already registered' } }, { status: 409 });
    }

    const passwordHash = await hashPassword(password);

    const user = await prisma.user.create({
      data: {
        fullName,
        email,
        phone,
        passwordHash,
        role: role as Role,
      }
    });

    const accessToken = await signAccessToken({ userId: user.id, email: user.email, role: user.role });
    const refreshToken = await signRefreshToken({ userId: user.id });

    return NextResponse.json({
      success: true,
      data: {
        user: { id: user.id, fullName: user.fullName, email: user.email, phone: user.phone, role: user.role },
        accessToken,
        refreshToken
      },
      message: 'Account created successfully'
    }, { status: 201 });

  } catch (error: unknown) {
    console.error('Registration Error:', error);
    return NextResponse.json({ success: false, error: { code: 'INTERNAL_ERROR', message: 'Something went wrong' } }, { status: 500 });
  }
}