# src/app/api/bookings/[id]/cancel/route.ts
import { NextResponse } from 'next/server';
import { prisma } from '@/lib/prisma';
import { verifyAccessToken } from '@/lib/auth';

export async function POST(req: Request, { params }: { params: Promise<{ id: string }> }) {
  try {
    const authHeader = req.headers.get('authorization')!;
    const token = authHeader.split(' ')[1];
    const payload = await verifyAccessToken(token);

    if (!payload) {
       return NextResponse.json({ success: false, error: { code: 'UNAUTHORIZED', message: 'Unauthorized' } }, { status: 401 });
    }

    const { id } = await params;
    const booking = await prisma.booking.findUnique({
      where: { id },
      include: { hotel: true }
    });

    if (!booking) {
      return NextResponse.json({ success: false, error: { code: 'NOT_FOUND', message: 'Booking not found' } }, { status: 404 });
    }

    // Only Booker who made it, or Hotel Manager of the hotel, or Admin can cancel
    if (booking.userId !== payload.userId && booking.hotel.managerId !== payload.userId && payload.role !== 'ADMIN') {
      return NextResponse.json({ success: false, error: { code: 'FORBIDDEN', message: 'Not authorized to cancel this booking' } }, { status: 403 });
    }

    if (booking.status !== 'PENDING') {
      return NextResponse.json({ success: false, error: { code: 'VALIDATION_ERROR', message: 'Only PENDING bookings can be cancelled directly' } }, { status: 400 });
    }

    const body = await req.json().catch(() => ({}));
    
    await prisma.booking.update({
      where: { id },
      data: {
        status: 'CANCELLED',
        cancelledAt: new Date(),
        cancelReason: body.reason || 'Cancelled by user'
      }
    });

    return NextResponse.json({ success: true, message: 'Booking cancelled' });
  } catch (error: unknown) {
    return NextResponse.json({ success: false, error: { code: 'INTERNAL_ERROR', message: 'Something went wrong' } }, { status: 500 });
  }
}